Move out numeric monitoring plugin interface to separate module
